+++ services/web-app/components/mdx/components.js
@@ -32,6 +32,7 @@ import ColorPalette from '@/components/color-palette'
 import ColorTokenTable from '@/components/color-token-table'
 import Divider from '@/components/divider'
 import { DoDont, DoDontRow } from '@/components/do-dont'
+import FilterableDesignKitTable from '@/components/filterable-design-kit-table'
 import FullPageError from '@/components/full-page-error'
 import GifPlayer from '@/components/gif-player'
 import Glossary from '@/components/glossary'
@@ -107,6 +108,7 @@ const components = {
   DoDont,
   DoDontRow,
   Divider,
+  FilterableDesignKitTable,
   GifPlayer,
   Glossary,
   Grid,